[task-6] implemented saving interested movies to list using Room

+++ data/src/main/java/com/example/data/apimodels/movies/Result.kt
@@ -1,5 +1,5 @@
 package com.example.data.apimodels.movies
 
 data class Result(
-    val id: Long
+    val id: Int
 )
